[DataTable] updates the selectable and sortable to be HoC to keep the specific data separated from the main table component
