+++ src/DataTable/TableHeader.js
@@ -2,46 +2,38 @@ import React, { PropTypes } from 'react';
 import classNames from 'classnames';
 import Tooltip from '../Tooltip';
 
-class TableHeader extends React.Component {
-    static propTypes = {
-        cellFormatter: PropTypes.func,
-        className: PropTypes.string,
-        name: PropTypes.string.isRequired,
-        numeric: PropTypes.bool,
-        onClick: PropTypes.func,
-        nosort: PropTypes.bool,
-        sortFn: PropTypes.func,
-        tooltip: PropTypes.node
-    };
+const propTypes = {
+    cellFormatter: PropTypes.func,
+    className: PropTypes.string,
+    name: PropTypes.string.isRequired,
+    numeric: PropTypes.bool,
+    onClick: PropTypes.func,
+    nosort: PropTypes.bool,
+    sortFn: PropTypes.func,
+    tooltip: PropTypes.node
+};
 
-    constructor(props) {
-        super(props);
+const TableHeader = props => {
+    const { className, name, numeric, onClick,
+        nosort, tooltip, children, ...otherProps } = props;
 
-        this.handleClick = this.handleClick.bind(this);
-    }
+    const classes = classNames({
+        'mdl-data-table__cell--non-numeric': !numeric
+    }, className);
 
-    handleClick(e) {
-        this.props.onClick(e, this.props.name);
-    }
+    const clickFn = !nosort && onClick
+        ? (e) => onClick(e, name)
+        : null;
 
-    render() {
-        const { cellFormatter, className, name, numeric, onClick,
-            nosort, sortFn, tooltip, children, ...otherProps } = this.props;
+    return (
+        <th className={classes} onClick={clickFn} {...otherProps}>
+            {!!tooltip ? (
+                <Tooltip label={tooltip}>{children}</Tooltip>
+            ) : children}
+        </th>
+    );
+};
 
-        const classes = classNames({
-            'mdl-data-table__cell--non-numeric': !numeric
-        }, className);
-
-        const clickFn = !nosort && onClick ? this.handleClick : null;
-
-        return (
-            <th className={classes} onClick={clickFn} {...otherProps}>
-                {!!tooltip ? (
-                    <Tooltip label={tooltip}>{children}</Tooltip>
-                ) : children}
-            </th>
-        );
-    }
-}
+TableHeader.propTypes = propTypes;
 
 export default TableHeader;
