+++ src/DataTable/index.js
@@ -1,4 +1,3 @@
 export TableHeader from './TableHeader';
 export Table from './Table';
-export SortableTable from './SortableTable';
 export default from './Table';
